Fix the mirror action example
We used the master branch of action/checkout. The master branch now does a shallow clone which cannot be mirrored ("shallow update not allowed").

We switch to the `v2-beta` branch. It still does a shallow clone but we can configured to fetch the whole history.

Also in the `v2-beta` branch no longer checks out the code in a detached state when specifying a branch name so we can now get rid of the checkout "hack".